Extract common Button / AboutDialog / GpsPosition defaults into BaseButton
