--- src/Mapbender/ManagerBundle/Resources/views/Element/about_dialog.html.twig
@@ -1,5 +1,0 @@
-<div class="elementForm">
-  {{ form_row(form.configuration.label) }}
-  {{ form_row(form.title) }}
-  {{ form_rest(form.configuration) }}
-</div>
